crud 수정, 권한정보에 따라 crud 제어

+++ src/main/java/aticle/articlev2/api/ArticleApiController.java
@@ -5,7 +5,9 @@ import aticle.articlev2.entity.Article;
 import aticle.articlev2.service.ArticleService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 
 import java.security.Principal;
@@ -42,8 +44,10 @@ public class ArticleApiController {
 
 
     //생성
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/articles/new")
     public ResponseEntity<ArticleDto> create(@RequestBody ArticleDto dto) {
+
         ArticleDto createdDto = articleService.create(dto);
 
         return ResponseEntity.status(HttpStatus.OK).body(createdDto);
@@ -55,19 +59,28 @@ public class ArticleApiController {
                                              @RequestBody ArticleDto dto,
                                              Principal principal) {
 
+        Article article = articleService.getArticle(id);
+        if (!article.getMember().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다");
+        }
+
         ArticleDto updated = articleService.update(id, dto);
-        return (updated != null) ?
-                ResponseEntity.status(HttpStatus.OK).body(updated) :
-                ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        if (updated != null) return ResponseEntity.status(HttpStatus.OK).body(updated);
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
     }
 
 
     //삭제
+    @PreAuthorize("isAuthenticated()")
     @DeleteMapping("/articles/{id}")
-    public ResponseEntity<Article> delete(@PathVariable Long id) {
+    public ResponseEntity<Article> delete(@PathVariable Long id,Principal principal) {
+        Article article = articleService.getArticle(id);
+        if (!article.getMember().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제 권한이 없습니다.");
+        }
+
         Article deleted = articleService.delete(id);
-        return (deleted != null) ?
-                ResponseEntity.status(HttpStatus.NO_CONTENT).build() :
-                ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        if (deleted != null) return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
     }
 }
